Add text frame layout and optional text structure

+++ src/Font/CidToGidMap.php
@@ -41,7 +41,7 @@ final class CidToGidMap extends IndirectObject
         }
 
         $maxCid = max(array_map(
-            static fn (string $code): int => hexdec($code),
+            static fn (string $code): int => (int) hexdec($code),
             array_keys($codeMap),
         ));
